requirements.txt
flask
langchain
langchain-community
langchain-core
chromadb
sentence-transformers
ollama
python-dotenv
gunicorn